Rename rundirs.py to computed_diagnostics.py
